@@ -43,9 +43,9 @@
         "i18next-xhr-backend": "^3.2.2",
         "lodash": "^4.17.21",
         "process": "^0.11.10",
-        "react": "^16.12.0",
+        "react": "^18.2.0",
         "react-apollo": "^3.1.5",
-        "react-dom": "^16.12.0",
+        "react-dom": "^18.2.0",
         "react-i18next": "^11.2.7",
         "react-redux": "^7.1.3",
         "react-router": "^5.1.2",
@@ -94,7 +94,7 @@
         "file-loader": "^3.0.1",
         "husky": "^3.0.5",
         "jest": "^28.1.0",
-        "jest-environment-jsdom": "^28.1.0",
+        "jest-environment-jsdom": "^29.5.0",
         "rimraf": "^2.6.2",
         "sass": "^1.52.1",
         "source-map-loader": "^1.0.0",
